requirements.txt
autopep8==2.0.2
boto3==1.26.50
botocore==1.29.50
jmespath==1.0.1
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.14
